+++ app/src/main/res/layout/activity_select_user_type.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:background="@drawable/background_gradient"
-    tools:context=".select_user_type">
+    tools:context=".SelectUserType">
 
     <!-- The primary full-screen view. This can be replaced with whatever view
          is needed to present your content, e.g. VideoView, SurfaceView,
@@ -14,14 +14,15 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:gravity="center"
+        android:gravity="center_horizontal"
+        android:layout_marginTop="16dp"
         android:id="@+id/app_details"
         >
 
         <ImageView
             android:id="@+id/iv_logo"
-            android:layout_width="150dp"
-            android:layout_height="150dp"
+            android:layout_width="200dp"
+            android:layout_height="230dp"
             android:background="@drawable/logo"
             />
